use PreviewConnectionScreen
